requirements.txt
fastapi
uvicorn[standard]
apscheduler
google-api-python-client
google-auth
google-auth-httplib2
pandas
openpyxl
python-dotenv
requests
